+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.109.0] - Not released
+### Fixed
+- Ignore minor exif-errors while sanitizing images.
+- Skip files which can not be sanitized because if errors after and report the errors.
 
 ## [1.108.1] - 2022-05-03
 ### Changed